Highlight :tags
Sadly, current implementation appears order dependent and doesn't allow to add
such support in user code or extensions, so we are forced to tweak the racket-mode
code instead. IMO the entire highlighting mechanism and its customization should
be exposed for the user to modify as needed but that requires upstream changes.